Port startup and exit routines
This change adapts various CRT routines to be capability-aware.
This includes routines to setup the initial stack, including converting
elements of argv etc to capabilities, porting memcpy and memset
implementations (the efficient version of strlen has to be disabled
for Morello as it may read past the end of a string), and some changes
to auxv-related code which is affected.

The definitions in bits appear to already be consistent with Morello,
however there are a couple of structs which in time will want to be
changed to accommodate 128-bit registers. I believe the structs in
user.h are only really intended for use by GDB, so are not needed to
get a minimum viable product off the ground. Similarly, the ones
in signal.h are for use by ucontext, which is deprecated. They should
both be fixed eventually.

+++ include/sys/auxv.h
@@ -8,7 +8,7 @@ extern "C" {
 #include <elf.h>
 #include <bits/hwcap.h>
 
-unsigned long getauxval(unsigned long);
+uintptr_t getauxval(unsigned long);
 
 #ifdef __cplusplus
 }
